:recycle: create HexStringWrapper and use it as Contribution aggregate ID
